Updated for 1.3
