+++ Hire/ChangeLog.txt
@@ -1,0 +1,10 @@
+0.1.0
+		Initial rebuild for KSP 1.3
+		Added standardized build scripts
+		Added AssemblyVersion.tt
+		Added .version file
+
+0.1.1	
+		Broke single file out into several files
+		Added Log.cs
+		Changed all logging to using standardized Log class
