+++ Hire/TRP-Hire.version
@@ -4,7 +4,7 @@
   "VERSION": {
     "MAJOR": 0,
     "MINOR": 1,
-    "PATCH": 0,
+    "PATCH": 1,
     "BUILD": 0
   },
   "KSP_VERSION": {
