+++ Hire/buildRelease.bat
@@ -39,5 +39,3 @@ echo Version:  %VERSION%
 set FILE="%RELEASEDIR%\%GAMEDIR%-%VERSION%.zip"
 IF EXIST %FILE% del /F %FILE%
 %ZIP% a -tzip %FILE% GameData
-
-pause
